+++ src/provider/interface.ts
@@ -18,6 +18,7 @@ import type {
   InvocationBulk,
   InvocationsDetailsWithNonce,
   InvokeFunctionResponse,
+  StateUpdateResponse,
   TransactionSimulationResponse,
   waitForTransactionOptions,
 } from '../types';
@@ -315,4 +316,12 @@ export abstract class ProviderInterface {
     invocationDetails: InvocationsDetailsWithNonce,
     blockIdentifier?: BlockIdentifier
   ): Promise<TransactionSimulationResponse>;
+
+  /**
+   * Gets the state changes in a specific block
+   *
+   * @param blockIdentifier - block identifier
+   * @returns StateUpdateResponse
+   */
+  public abstract getStateUpdate(blockIdentifier?: BlockIdentifier): Promise<StateUpdateResponse>;
 }