+++ src/provider/utils.ts
@@ -1,5 +1,5 @@
 /* eslint-disable max-classes-per-file */
-import type { BlockNumber } from '../types';
+import type { BlockNumber, SequencerIdentifier } from '../types';
 import { BigNumberish, isHex, toHex } from '../utils/number';
 
 /**
@@ -95,4 +95,10 @@ export class Block {
   valueOf = () => this.number;
 
   toString = () => this.hash;
+
+  get sequencerIdentifier(): SequencerIdentifier {
+    return this.hash !== null
+      ? { blockHash: this.hash as string }
+      : { blockNumber: (this.number ?? this.tag) as BlockNumber };
+  }
 }