+++ src/utils/uint256.ts
@@ -22,7 +22,7 @@ export function isUint256(bn: BigNumberish): boolean {
   return toBigInt(bn) <= UINT_256_MAX;
 }
 
-// function to convert BN to Uint256
+// function to convert BN to Uint256 hex
 export function bnToUint256(bignumber: BigNumberish): Uint256 {
   const bn = toBigInt(bignumber);
   if (!isUint256(bn)) throw new Error('Number is too large');